requirements.txt
matplotlib>=3.1
nibabel>=3.0.0
numpy>=1.18
pandas>=0.25
scikit-image>=0.15
tqdm>=4.38
lightning
torchvision
tensorboard